+++ actioncable/app/assets/javascripts/action_cable.js
@@ -28,6 +28,22 @@
       throw new TypeError("Cannot call a class as a function");
     }
   };
+  var createClass = function() {
+    function defineProperties(target, props) {
+      for (var i = 0; i < props.length; i++) {
+        var descriptor = props[i];
+        descriptor.enumerable = descriptor.enumerable || false;
+        descriptor.configurable = true;
+        if ("value" in descriptor) descriptor.writable = true;
+        Object.defineProperty(target, descriptor.key, descriptor);
+      }
+    }
+    return function(Constructor, protoProps, staticProps) {
+      if (protoProps) defineProperties(Constructor.prototype, protoProps);
+      if (staticProps) defineProperties(Constructor, staticProps);
+      return Constructor;
+    };
+  }();
   var now = function now() {
     return new Date().getTime();
   };
@@ -432,7 +448,7 @@
   var Consumer = function() {
     function Consumer(url) {
       classCallCheck(this, Consumer);
-      this.url = url;
+      this._url = url;
       this.subscriptions = new Subscriptions(this);
       this.connection = new Connection(this);
     }
@@ -452,28 +468,36 @@
         return this.connection.open();
       }
     };
+    createClass(Consumer, [ {
+      key: "url",
+      get: function get$$1() {
+        return createWebSocketURL(this._url);
+      }
+    } ]);
     return Consumer;
   }();
+  function createWebSocketURL(url) {
+    if (typeof url === "function") {
+      url = url();
+    }
+    if (url && !/^wss?:/i.test(url)) {
+      var a = document.createElement("a");
+      a.href = url;
+      a.href = a.href;
+      a.protocol = a.protocol.replace("http", "ws");
+      return a.href;
+    } else {
+      return url;
+    }
+  }
   function createConsumer() {
     var url = arguments.length > 0 && arguments[0] !== undefined ? arguments[0] : getConfig("url") || INTERNAL.default_mount_path;
-    return new Consumer(createWebSocketURL(url));
+    return new Consumer(url);
   }
   function getConfig(name) {
     var element = document.head.querySelector("meta[name='action-cable-" + name + "']");
     if (element) {
       return element.getAttribute("content");
-    }
-  }
-  function createWebSocketURL(url) {
-    var webSocketURL = typeof url === "function" ? url() : url;
-    if (webSocketURL && !/^wss?:/i.test(webSocketURL)) {
-      var a = document.createElement("a");
-      a.href = webSocketURL;
-      a.href = a.href;
-      a.protocol = a.protocol.replace("http", "ws");
-      return a.href;
-    } else {
-      return webSocketURL;
     }
   }
   exports.Connection = Connection;
@@ -483,10 +507,10 @@
   exports.Subscription = Subscription;
   exports.Subscriptions = Subscriptions;
   exports.adapters = adapters;
+  exports.createWebSocketURL = createWebSocketURL;
   exports.logger = logger;
   exports.createConsumer = createConsumer;
   exports.getConfig = getConfig;
-  exports.createWebSocketURL = createWebSocketURL;
   Object.defineProperty(exports, "__esModule", {
     value: true
   });