+++ actioncable/test/javascript/src/unit/action_cable_test.js
@@ -42,12 +42,16 @@ module("ActionCable", () => {
       assert.equal(consumer.url, testURL)
     })
 
-    test("uses function to generate URL", assert => {
+    test("dynamically computes URL from function", assert => {
+      let dynamicURL = testURL
       const generateURL = () => {
-        return testURL
+        return dynamicURL
       }
       const consumer = ActionCable.createConsumer(generateURL)
       assert.equal(consumer.url, testURL)
+
+      dynamicURL = `${testURL}foo`
+      assert.equal(consumer.url, `${testURL}foo`)
     })
   })
 })