+++ actionview/lib/action_view/renderer/template_renderer.rb
@@ -26,7 +26,12 @@ module ActionView
         elsif options.key?(:html)
           Template::HTML.new(options[:html], formats.first)
         elsif options.key?(:file)
-          @lookup_context.with_fallbacks.find_file(options[:file], nil, false, keys, @details)
+          if File.exist?(options[:file])
+            Template::RawFile.new(options[:file])
+          else
+            ActiveSupport::Deprecation.warn "render file: should be given the absolute path to a file"
+            @lookup_context.with_fallbacks.find_file(options[:file], nil, false, keys, @details)
+          end
         elsif options.key?(:inline)
           handler = Template.handler_for_extension(options[:type] || "erb")
           format = if handler.respond_to?(:default_format)
@@ -84,6 +89,7 @@ module ActionView
         when String
           begin
             if layout.start_with?("/")
+              ActiveSupport::Deprecation.warn "Rendering layouts from an absolute path is deprecated."
               @lookup_context.with_fallbacks.find_template(layout, nil, false, [], details)
             else
               @lookup_context.find_template(layout, nil, false, [], details)